requirements.txt
einops==0.6.1
# h5py==3.5.0
icecream==2.1.3
imageio==2.10.1
MedPy==0.4.0
opencv_python==4.5.4.58
pycocotools==2.0.6
safetensors==0.3.1
scipy==1.7.3
SimpleITK==2.2.1
tensorboardX==2.6
# torch==1.13.1 # maybe change to torch 2?
# torchvision==0.10.1+cu111
tqdm==4.62.3
ml-collections==0.1.1
pycocotools==2.0.6
python-dotenv
dill
# onnx==1.13.1
# onnxruntime==1.14.1
monai
elasticdeform
vedo
opencv-python-headless
pandas
matplotlib
tensorboard
nibabel
deepdiff
numpy